students/K33402/Matrokhina_Anna/LR3/warriors/requirements.txt
django==3.2.11
djangorestframework
markdown
django-filter
drf-yasg